Add New Components and Refactor Structure for Enhanced Modularity
- Introduced multiple new components including BlogHeader, BlogPost, CodeBlock, Table, Tooltip, and various UI elements to improve code organization and reusability.
- Refactored existing components to align with the new structure, ensuring a clear separation of concerns and adherence to best practices in TypeScript and Svelte development.
- Updated routes to reflect the new component paths, enhancing maintainability and clarity in the project structure.
- Added unit tests for new components to ensure functionality and reliability, following testing guidelines for comprehensive coverage.
