+++ src/lib/components/blog/BlogPost.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import Box from '$lib/components/Box.svelte';
+	import Box from '$lib/components/layout/Box.svelte';
+	import type { BlogPost } from '$lib/types';
 	import { formatDate } from '$services/blog-service';
 
 	const { post } = $props<{
