+++ src/lib/components/projects/GoProjectSetup.svelte
