+++ src/lib/components/projects/ProjectCard.svelte
