+++ src/lib/components/projects/ProjectGrid.svelte
