+++ src/lib/components/resources/ResourceCard.svelte
