+++ src/lib/components/resources/ResourceHeader.svelte
