+++ src/lib/components/resources/ResourceSection.svelte
