+++ src/lib/components/terminal/Terminal.svelte
