+++ src/lib/components/terminal/Terminal.test.ts
