+++ src/lib/components/ui/Badge.svelte
