+++ src/lib/components/ui/Footer.svelte
