+++ src/lib/components/ui/Footer.test.ts
