+++ src/routes/blog/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { blogPosts, fetchFeed } from '$services/blog-service';
 	import { onMount } from 'svelte';
-	import BlogHeader from '$lib/components/BlogHeader.svelte';
-	import BlogPost from '$lib/components/BlogPost.svelte';
+	import BlogHeader from '$lib/components/blog/BlogHeader.svelte';
+	import BlogPost from '$lib/components/blog/BlogPost.svelte';
 
 	let currentPage = $state(1);
 	let hasMore = $state(false);
